gnu: shepherd: Build Shepherd from git.
The rationale for this is that it makes it simpler to start VM's with a
Shepherd that is compiled from a local checkout.  This facilitates hacking on
Shepherd itself, and testing it in the Guix init environment, by speeding up
the edit-compile-test cycle.

* gnu/packages/admin.scm (shepherd-0.10): Source points to the git repo; adjust
native-inputs accordingly.  Also delete the strip build phase.